refactor state into useReducer hook
