+++ wp-content/themes/labor-day/src/js/search/Model.ts
@@ -1,4 +1,4 @@
-import { POSTS_PER_PAGE, graphQL } from './Utilities';
+import { POSTS_PER_PAGE, graphQL } from './Utilities/Utilities';
 import { makeRequestResponse, queryVars } from './types/promises';
 
 export default new (class Model {
