+++ wp-content/themes/labor-day/src/js/search/Presentational/Posts/FeaturedImage.tsx
@@ -1,5 +1,9 @@
+// 3rd Party
 import React from '@wordpress/element';
-import { featuredImage } from '../types';
+
+// Types
+import { featuredImage } from '../../types';
+
 export default function FeaturedImage({
 	featuredImage,
 }: {
