+++ wp-content/themes/labor-day/src/js/search/Presentational/Search Bar/ResultsContainer.tsx
@@ -1,0 +1,82 @@
+// 3rd Party
+import React from '@wordpress/element';
+
+// Types
+import { PrettyEventData } from '../../types';
+import { selectedFilterObject } from '../../types/eventFilters';
+
+// Components
+import SinglePost from '../Posts/SinglePost';
+import { ReadMoreButton } from '../CNOButtons';
+import { AddToScheduleButton } from '../../Components/AddToScheduleButton';
+
+// Utilities
+import { createExcerpt } from '../../Utilities/Utilities';
+
+// The Component
+export default function ResultsContainer({
+	posts,
+	dispatch,
+	selectedFilters,
+}: {
+	posts: PrettyEventData[];
+	selectedFilters: selectedFilterObject;
+	dispatch: Function;
+}) {
+	const emptyFilters = Object.values(selectedFilters).every(
+		(filter) => 'Select Option' === filter
+	);
+	return (
+		<section className="cno-events">
+			{posts.map((post: PrettyEventData) => {
+				if (emptyFilters || postMatchesFilters(post, selectedFilters)) {
+					return (
+						<SinglePost
+							data={post}
+							dispatch={dispatch}
+							key={post.eventId}
+						>
+							<>
+								<AddToScheduleButton eventId={post.eventId} />
+								{createExcerpt(post.event_info.description)
+									.readMore && (
+									<ReadMoreButton link={post.link} />
+								)}
+							</>
+						</SinglePost>
+					);
+				}
+				return null;
+			})}
+		</section>
+	);
+}
+
+/** Returns true if Post matches selected filters
+ * @param {PrettyEventData} post The post to check
+ * @returns boolean
+ */
+function postMatchesFilters(
+	post: PrettyEventData,
+	selectedFilters: selectedFilterObject
+): boolean {
+	const postInFilters: boolean =
+		(filterIsEmpty(selectedFilters.Locations) ||
+			post.locations?.some(
+				(location) => selectedFilters.Locations === location.name
+			)) &&
+		(filterIsEmpty(selectedFilters['Event Types']) ||
+			post.type?.some(
+				(type) => selectedFilters['Event Types'] === type.name
+			)) &&
+		(filterIsEmpty(selectedFilters.Days) ||
+			selectedFilters.Days === post.event_info.info.day);
+	return postInFilters;
+}
+/** Returns true if filter is default "empty" value
+ * @param {string} filter the filter to check
+ * @returns boolean
+ */
+function filterIsEmpty(filter: string): boolean {
+	return filter === 'Select Option';
+}
