+++ wp-content/themes/labor-day/src/js/search/Presentational/Search Bar/SearchBarContainer.jsx
@@ -1,0 +1,12 @@
+import React from '@wordpress/element';
+
+export default function SearchBarContainer({ children }) {
+	return (
+		<section className="cno-event-search">
+			<div className="container">
+				<h2 className="cno-event-search__title">Search Events</h2>
+				{children}
+			</div>
+		</section>
+	);
+}
