--- wp-content/themes/labor-day/src/js/search/Presentational/SearchInput.jsx
@@ -1,13 +1,0 @@
-export default function SearchInput( { search, handleSearchInput } ) {
-	return (
-		<input
-			type="text"
-			name="search"
-			id="search"
-			placeholder="Find an Event"
-			className="cno-event-search__search-bar"
-			value={ search }
-			onChange={ handleSearchInput }
-		/>
-	);
-}
