+++ wp-content/themes/labor-day/src/js/search/Presentational/ShareModal.tsx
@@ -4,10 +4,11 @@ declare const cnoSiteData: { rootUrl: string };
 
 export default function ShareModal({
 	showShareModal,
-	setShowShareModal,
 	shareEventObject,
+	dispatch,
 }) {
-	const url = encodeURIComponent(shareEventObject.link);
+	const { link, title } = shareEventObject;
+	const url = encodeURIComponent(link);
 	return (
 		<div
 			style={{
@@ -16,7 +17,8 @@ export default function ShareModal({
 			}}
 			className="share-event__overlay overlay"
 			onClick={(ev) => {
-				setShowShareModal(false);
+				if (!ev.target.classList.contains('overlay')) return;
+				dispatch({ type: 'closeModal' });
 			}}
 		>
 			<dialog
@@ -27,7 +29,7 @@ export default function ShareModal({
 				<ul className="share-locations">
 					<li>
 						<a
-							href={`https://www.facebook.com/dialog/share?app_id=3353870511533012&display=popup&href=${shareEventObject.link}&redirect_uri=${cnoSiteData.rootUrl}/events`}
+							href={`https://www.facebook.com/dialog/share?app_id=3353870511533012&display=popup&href=${link}&redirect_uri=${cnoSiteData.rootUrl}/events`}
 							title="Share on Facebook"
 							target="_blank"
 							className="share-locations__location--facebook btn__outline--secondary"
@@ -38,7 +40,7 @@ export default function ShareModal({
 					</li>
 					<li>
 						<a
-							href={`sms:?&body=Halito%20(Hello)! I'm going to "${shareEventObject.title}" and I thought you'd like to check it out, too! Learn more at ${url}`}
+							href={`sms:?&body=Halito%20(Hello)! I'm going to "${title}" and I thought you'd like to check it out, too! Learn more at ${url}`}
 							target="_blank"
 							className="share-locations__location--text btn__outline--secondary"
 						>
@@ -48,7 +50,7 @@ export default function ShareModal({
 					</li>
 					<li>
 						<a
-							href={`mailto:?subject=Check out this event!&body=Halito%20(Hello)! I'm going to "${shareEventObject.title}" and I thought you'd like to check it out, too! Learn more at ${url}.`}
+							href={`mailto:?subject=Check out this event!&body=Halito%20(Hello)! I'm going to "${title}" and I thought you'd like to check it out, too! Learn more at ${url}.`}
 							title="Share via Email"
 							className="share-locations__location--email btn__outline--secondary"
 						>
